+++ packages/nodes-base/nodes/EmailSend/EmailSend.node.json
@@ -29,5 +29,6 @@
 	},
 	"subcategories": {
 		"Core Nodes": ["Helpers"]
-	}
+	},
+	"alias": ["SMTP"]
 }
