+++ packages/nodes-base/nodes/EmailSend/v2/EmailSendV2.node.ts
@@ -1,0 +1,83 @@
+/* eslint-disable n8n-nodes-base/node-filename-against-convention */
+import { IExecuteFunctions } from 'n8n-core';
+
+import {
+	INodeExecutionData,
+	INodeType,
+	INodeTypeBaseDescription,
+	INodeTypeDescription,
+} from 'n8n-workflow';
+
+import * as send from './send.operation';
+
+const versionDescription: INodeTypeDescription = {
+	displayName: 'Send Email',
+	name: 'emailSend',
+	icon: 'fa:envelope',
+	group: ['output'],
+	version: 2,
+	subtitle: '={{ $parameter["operation"] + ": " + $parameter["resource"] }}',
+	description: 'Sends an Email',
+	defaults: {
+		name: 'Send Email',
+		color: '#00bb88',
+	},
+	inputs: ['main'],
+	outputs: ['main'],
+	credentials: [
+		{
+			name: 'smtp',
+			required: true,
+		},
+	],
+	properties: [
+		{
+			displayName: 'Resource',
+			name: 'resource',
+			type: 'hidden',
+			noDataExpression: true,
+			default: 'email',
+			options: [
+				{
+					name: 'Email',
+					value: 'email',
+				},
+			],
+		},
+		{
+			displayName: 'Operation',
+			name: 'operation',
+			type: 'hidden',
+			noDataExpression: true,
+			default: 'send',
+			options: [
+				{
+					name: 'Send',
+					value: 'send',
+				},
+			],
+		},
+		...send.description,
+	],
+};
+
+export class EmailSendV2 implements INodeType {
+	description: INodeTypeDescription;
+
+	constructor(baseDescription: INodeTypeBaseDescription) {
+		this.description = {
+			...baseDescription,
+			...versionDescription,
+		};
+	}
+
+	async execute(this: IExecuteFunctions): Promise<INodeExecutionData[][]> {
+		const operation = this.getNodeParameter('operation', 0);
+		let returnData: INodeExecutionData[][] = [];
+
+		if (operation === 'send') {
+			returnData = await send.execute.call(this);
+		}
+		return returnData;
+	}
+}
