+++ packages/nodes-base/utils/allCurrencies.ts
